Tons of cleanup
-Restructuring and cleanup of major site components
-Header restructure and visual separation
-devlog experimentation
