+++ components/layout.tsx
@@ -1,11 +1,11 @@
 import Head from "next/head";
 import Image from "next/image";
 import styles from "./layout.module.css";
-import utilStyles from "../styles/utils.module.css";
 import Link from "next/link";
 import bg from "../public/images/parkslope.jpg";
 import Providers from "./providers";
 import Header from "../sections/Header"
+import Footer from "../sections/Footer"
 
 const name = "Cooper Reed";
 export const siteTitle = "Co-Operability";
@@ -40,6 +40,7 @@ export default function Layout({ children, home }) {
           </div>
         )}
       </Providers>
+      <Footer />
     </div>
   );
 }
