+++ pages/devlog.tsx
@@ -2,7 +2,6 @@ import Link from "next/link";
 import Date from "../components/date";
 import utilStyles from "../styles/utils.module.css";
 import Layout, { siteTitle } from "../components/layout";
-import Footer from "../sections/Footer";
 import { getSortedPostsData } from "../lib/posts";
 
 export default function DevLog({ allPostsData }) {
@@ -24,7 +23,6 @@ export default function DevLog({ allPostsData }) {
                     ))}
                 </ul>
             </section>
-            <Footer />
         </Layout>
     );
 };
