+++ posts/2023-08-21.md
@@ -1,0 +1,18 @@
+---
+title: "Markdown Cheat Sheet"
+date: "2023-08-21"
+---
+
+New links are up:
+**[cooperreed.me][1]** and **[cooperability.com][2]** both redirect to cooperability.vercel.app.
+
+```
+Code
+```
+
+Yours,
+
+- **Cooper**
+
+  [1]: https://cooperreed.me/ "link 1"
+  [2]: https://cooperability.com/ "link 2"
