+++ styles/global.css
@@ -28,7 +28,7 @@ img {
 }
 
 button {
-  color:#0070f3;
+  color: #0070f3;
   text-decoration: underline;
 }
 
@@ -38,17 +38,18 @@ button {
 
 @layer base {
   body {
-    @apply bg-gray-50 text-gray-900 dark:bg-black dark:text-gray-100;
-    @apply min-h-screen;
-    body {
-    color: #000;
-    background-color: #fff;
-}
+    /* @apply bg-gray-50 text-gray-900 dark:bg-black dark:text-gray-100; */
+    /* @apply min-h-screen; */
 
-[data-theme="dark"] body {
-    color:#fff;
-    background-color: #000;
-}
+    body {
+      color: #000;
+      background-color: #fff;
+    }
+
+    [data-theme="dark"] body {
+      color: #fff;
+      background-color: black;
+    }
   }
 
   ::selection {
@@ -59,4 +60,8 @@ button {
   .btn {
     @apply inline-block px-5 py-3 rounded-lg hover: bg-indigo-600 transition duration-300 shadow-lg bg-indigo-500 text-white capitalize tracking-wider font-semibold text-sm sm:text-xl focus:outline-none focus:ring-4 focus:ring-indigo-500 focus:ring-opacity-50 whitespace-nowrap
   }
+}
+
+@layer components {
+  
 }
