+++ styles/utils.module.css
@@ -71,7 +71,6 @@
   padding: 0;
   margin: 0;
   z-index: 1;
-
 }
 
 .listItem {
@@ -109,7 +108,7 @@
 
 .horizLine {
   border: 2px solid;
-  margin: 10px 5px;
+  margin: 10px 3px;
 }
 
 .hoverImage {
