+++ tsconfig.json
@@ -22,7 +22,7 @@
     "next-env.d.ts",
     "**/*.ts",
     "**/*.tsx"
-, "sections/Layout.tsx", "components/layout.tsx", "components/providers.js", "pages/quote.tsx", "pages/posts/[id].js", "postcss.config.js", "tailwind.config.js", "components/date.js"  ],
+, "components/layout.tsx", "components/providers.js", "pages/quote.tsx", "pages/posts/[id].js", "postcss.config.js", "tailwind.config.js", "components/date.js"  ],
   "exclude": [
     "node_modules"
   ]
